refactor: keep volume populator dependencies together
